implementation of services and communication completed
