+++ InnovativeProjectService/src/com/PurchaseServiceAPI.java
@@ -38,9 +38,17 @@ public class PurchaseServiceAPI {
 
 	public String sendDeatailsToPayment(int i_projectid, String pname, double pPrice) {
 
-		// create user JSON Object
+		
+		//  create CommunicationService class object
+		
+		CommunicationService comObj = new CommunicationService();
+		
+		// call the method and store in a string variable
 
-		String currentUserDetails = getCurrentLoggedUserinfo();
+		String currentUserDetails = comObj.getCurrentLoggedUserinfo();
+		
+		// create user JSON Object
+		
 		JsonObject userJSONobj = new JsonParser().parse(currentUserDetails).getAsJsonObject();
 		// get by user account communication
 
@@ -88,20 +96,6 @@ public class PurchaseServiceAPI {
 
 	}
 
-	public String getCurrentLoggedUserinfo() {
-
-		Client client = Client.create();
-
-		WebResource webResource = client.resource("http://localhost:8083/UserAccounts/UserAccountService/User_logins");
-
-		ClientResponse response = webResource.accept("application/json").get(ClientResponse.class);
-
-		String output = response.getEntity(String.class);
-
-		System.out.println(output);
-
-		return output;
-
-	}
+	
 
 }
